Combine drain spells
